+++ opencore_legacy_patcher/sys_patch/patchsets/hardware/graphics/amd_polaris.py
@@ -139,10 +139,9 @@ class AMDPolaris(BaseHardware):
                 **AMDOpenCL(self._xnu_major, self._xnu_minor, self._constants.detected_os_version).patches(),
             })
 
-        if self._is_gpu_architecture_present(gpu_architectures=[device_probe.Intel.Archs.Skylake]) is False:
-            # Monterey GVA is not required for Intel Skylake iGPUs
-            _base.update({
-                **MontereyGVA(self._xnu_major, self._xnu_minor, self._constants.detected_os_version).patches(),
-            })
+        # Polaris and Vega GPUs still use the native GVA stack
+        _base.update({
+            **MontereyGVA(self._xnu_major, self._xnu_minor, self._constants.detected_os_version).revert_patches(),
+        })
 
         return _base